+++ src/test/java/tests/US1008/TC01.java
@@ -1,9 +1,7 @@
 package tests.US1008;
 
 import org.testng.annotations.Test;
-import pages.US1008;
-import utilities.ConfigReader;
-import utilities.Driver;
+import pages.RoomReservations;
 
 public class TC01 {
 
@@ -13,10 +11,8 @@ public class TC01 {
     //Yönetici olarak Anasayfaya sorunsuz giriş yapabilmeliyim
     //Yönetici  sağ üst köşede "Log İn" sayfasını tıklamalıdır
     //Yönetici  username ve password girip "Log in" butonuna basmalıdır
-
-    US1008 us1008=new US1008();
-    us1008.girisYap();
-
+    RoomReservations roomReservations=new RoomReservations();
+    roomReservations.girisYap();
