+++ src/test/java/tests/US1008/TC03.java
@@ -1,7 +1,7 @@
 package tests.US1008;
 
 import org.testng.annotations.Test;
-import pages.US1008;
+import pages.RoomReservations;
 
 public class TC03 {
     //Kullanıcılar Anasayfaya sorunsuz erişebilmelidir.
@@ -12,10 +12,10 @@ public class TC03 {
 
     @Test
     public void roomReservation() {
-        US1008 us1008 = new US1008();
-        us1008.girisYap();
-        us1008.hotelManegementLinki.click();
-        us1008.RoomReservationLinki.click();
+        RoomReservations roomReservations=new RoomReservations();
+        roomReservations.girisYap();
+        roomReservations.hotelManegementLinki.click();
+        roomReservations.RoomReservationLinki.click();
 
     }
 }
