+++ src/test/java/tests/US1008/TC04.java
@@ -1,8 +1,7 @@
 package tests.US1008;
 
-import org.openqa.selenium.Keys;
 import org.testng.annotations.Test;
-import pages.US1008;
+import pages.RoomReservations;
 
 public class TC04 {
     //Kullanıcılar Anasayfaya sorunsuz erişebilmelidir
@@ -14,11 +13,11 @@ public class TC04 {
 
     @Test
     public void addRoomReservation() {
-        US1008 us1008 = new US1008();
-        us1008.girisYap();
-        us1008.hotelManegementLinki.click();
-        us1008.RoomReservationLinki.click();
-        us1008.AddRoomReservationLinki.click();
+        RoomReservations roomReservations=new RoomReservations();
+        roomReservations.girisYap();
+        roomReservations.hotelManegementLinki.click();
+        roomReservations.RoomReservationLinki.click();
+        roomReservations.AddRoomReservationLinki.click();
